Added stores (among other changes). You can buy from them, but not sell to them yet.
Added an initialize_world() function. In it, factions (and stores) are generated at the beginning of the game (so they can change throughout the game without it permanently changing it for other games).
Player classes can now determine how much money the creature starts with.
The mouse is now usable on the new game screen.
Added a "store" feature that opens th estore screen when the player steps on it.
Added an Item:clone() function that creates and returns a clone of the item.