Added a "multiselect" gamestate and actions attached to Features.
Multiselect:
The Multiselect dialog takes a list of entries. Each entry holds the information about what function they will call and what arguments they will pass to that function.
As an example, the spellbook item makes use of Multiselect to enable the player to select a spell to learn from the book.

Feature Actions:
Features can have actions attached to them, that trigger an action() callback on the feature.
When the player is next to a feature with an action, pressing the Action key will perform the action. If next to two features, it will pop up a Multiselect window asking the player to select which action they want.
Theoretically, features can have multiple actions attached to them at the same time, but this is untested.
The door feature now has opening the door as an action (though moving into a closed door also opens it), and close as an action if it's opened.