Added player species
Redid the newgame screen entirely to adjust for picking species and classes. Gamedefinition has settings to disable player species and classes, and to define the creature definition to use as the "default" player
Redid initialize_player to take a lot of (optional) arguments now, to set player name, determine which creature definition to use, set gender and class
Added creature and item density values to gamedefinition, branches and maptypes
Added generic leveling for creatures, creatures can generate at higher levels than their base level now
Added a Creature:get_stat() funciton that returns 0 if the stat in questio ndoesn't exist
Added font size options. Fixed a bunch of screens to display well with varying font sizes, but still need to fix some (namely the loading saves screen, stores and factions, multiselect, crafting)

A lot of this stuff still needs work, but might not get much chance to work on it until the end of the year so wanted to get in one last commit