Projectiles can now create effects and features when they hit (or miss) their target.
Effects and features created in this war have their "creator" field assigned to the creature who created the projectile, so that damage done by the effect/feature will be properly assigned.
Redid a bunch of effects and projectiles to refer to "creator" rather than "caster" or "owner" or whatever else they may have been, for consistency.